+++ cfgov/cfgov/settings/production.py
@@ -1,3 +1,1 @@
 from .base import *
-
-STAGING_HOSTNAME = 'content.consumerfinance.gov'