+++ cfgov/v1/templatetags/share.py
@@ -1,5 +1,7 @@
+import os
+
 from django import template
-from wagtail.wagtailcore.models import Page, Site
+from wagtail.wagtailcore.models import Page
 from django.conf import settings
 
 register = template.Library()
@@ -19,7 +21,8 @@ def is_shared(page):
 def staging_url(context, page):
     url = page.url
     if settings.STAGING_HOSTNAME not in page.url:
-        url = url.replace(context['request'].site.hostname, settings.STAGING_HOSTNAME)
+        url = url.replace(context['request'].site.hostname,
+                          os.environ.get('STAGING_HOSTNAME'))
     return url